+++ exonum-java-binding/service-archetype/src/main/resources/archetype-resources/src/main/java/MySchema.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package ${groupId};
+package ${package};
 
 import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.core.service.Schema;