+++ exonum-java-binding/service-archetype/src/main/resources/archetype-resources/src/main/java/MyTransactionConverter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package ${groupId};
+package ${package};
 
 import com.exonum.binding.core.transaction.Transaction;
 import com.exonum.binding.core.transaction.RawTransaction;